pyproject.toml
[tool.poetry]
name = "koi-data-manager"
version = "1.0.0"
description = "Streamlit web app that can take the details from the users. Finally give the Path for uploading the data on aws S3 bucket and meta_data.json file."
authors = ["KoiReader Technologies <support@koireader.com>"]
readme = "README.md"
license = "proprietary"

[tool.poetry.dependencies]
python = "^3.9.12"
streamlit = "^1.22.0"
pydantic = "^1.10.7"
nanoid = "^2.0.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.9"
streamlit-modal = "^0.1.0"
botocore = "^1.31.19"
boto3 = "^1.28.19"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
